Fix infinite loop in sample selection
When all relevant samplers are exhausted, the KLD sample selection will
give identically zero KLD for all k. This just means that within the
bounds of what we are "allowed" to do, there is nothing that would give
more information. In that case, the best we can do is to just abort and
roll with what we have; this corresponds to scenarios where there are
two equally good options (within the computational effort we are willing
to expend), so we essentially just flip a coin to decide between them.

Notably, these are the cases where the hierarchical Bayesian
approximation is worst; however, they also turn out to be rather rare
anyways.